# express-server/api-spec.yaml
openapi: 3.0.0
x-stoplight:
  id: hp9sgyu2laazz
info:
  title: event-calendar
  version: '1.0'
  description: Backend API for the EventCalendar App.
servers:
  - url: 'http://localhost:3000'
paths:
  /event:
    post:
      summary: Creates a new event
      operationId: delete-event
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventId:
                    type: string
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      description: |-
        Creates a new event.

        Returns the ID of the newly created event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      tags:
        - Event
    delete:
      summary: Deletes old events
      operationId: delete-event
      responses:
        '200':
          description: OK
      description: Deletes old events with past endDates
      tags:
        - Event
    get:
      summary: Gets all events
      operationId: get-event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      description: Gets all events
      tags:
        - Event
  '/event/{eventId}':
    parameters:
      - schema:
          type: string
        name: eventId
        in: path
        required: true
    get:
      summary: Gets an Event
      tags:
        - Event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '201':
          description: Created
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Event not found
      operationId: get-event-eventId
      description: Gets an Event with a given ID
      parameters: []
    patch:
      summary: Updates an event
      operationId: patch-event-eventId
      responses:
        '204':
          description: Event updated
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Event not found
      description: Updates an event with the given ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      tags:
        - Event
    delete:
      summary: Deletes an event
      operationId: delete-event-eventId
      responses:
        '204':
          description: Event deleted
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Event not found
      description: Deletes an event with the given ID
      tags:
        - Event
  '/event/{eventId}/assign':
    parameters:
      - schema:
          type: string
        name: eventId
        in: path
        required: true
    get:
      summary: Gets all the user's assigned to an event
      tags:
        - Event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Event not found
      operationId: get-event-eventId-assign
      description: Gets users assigned to an event
  '/event/{eventId}/assign/{userId}':
    parameters:
      - schema:
          type: string
        name: eventId
        in: path
        required: true
      - schema:
          type: string
        name: userId
        in: path
        required: true
    post:
      summary: Assign an event to a user
      tags:
        - User
        - Event
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendanceId:
                    type: string
            application/xml:
              schema:
                type: object
                properties:
                  attendanceId:
                    type: string
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Event or user not found
      operationId: get-event-eventId-assign-userId
      description: Assigns a event with a given ID to a user with a given ID. Returns the ID of the attendance of a user to an an event.
    delete:
      summary: Removes an Assignment
      operationId: delete-event-eventId-assign-userId
      responses:
        '200':
          description: OK
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: No assignment was found - so not deleted. Double check that eventId and userId actually reference an event and user
      description: Removes an assignment from a user to an event
      tags:
        - Event
        - User
  /user:
    post:
      summary: Creates a new user
      operationId: post-user
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      description: 'Creates a new user, returns an ID of the newly created user'
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                isAdmin:
                  type: string
                email:
                  type: string
                'password ':
                  type: string
    get:
      summary: Get all users
      operationId: get-user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      description: |-
        Gets all users

        By default only returns users that aren't admins, can be specified in query params if needed
      tags:
        - User
      parameters:
        - schema:
            type: boolean
          in: query
          name: includeAdmins
          description: Whether system administrators should be included in results or not. Default is false
  '/user/{userId}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: Gets a user
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: User not found
      operationId: get-user-userId
      description: Gets a user with the given ID
      tags:
        - User
    delete:
      summary: Deletes a user
      operationId: delete-user-userId
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: User not found
      description: |-
        Deletes a user with a given ID

        User must have necessary permissions
      tags:
        - User
    patch:
      summary: Updates a user
      operationId: patch-user-userId
      responses:
        '204':
          description: User successfully updated
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Not Found
      description: |-
        Updates a user

        User must have permissions in order to update.
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - firstName
                - lastName
                - email
                - password
  '/user/{userId}/events':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: Gets the events for a user
      tags:
        - User
        - Event
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: User not found
      operationId: get-user-userId-events
      description: Gets all the events that are assigned to a user with the given ID
  '/user/{userId}/photo':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: Retrieves the photo of a given user
      tags:
        - User
      responses:
        '200':
          description: |-
            User found

            Contains image as jpg
          content:
            image/jpeg:
              schema:
                type: object
                properties: {}
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          description: Photo not found
      operationId: get-user-userOId
      description: Finds the photo of a user with the given ID
    post:
      summary: Upload or Update Photo
      operationId: post-user-userId-photo
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      description: Updates or creates a photo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
        description: |-
          A multipart form, with just one entry for the photo.

          Photo should have key 'photo' and be jpg. 

          I don't know how to document photos with stoplight, so I'll just describe how to interact with this endpoint.

          Basically the multi part form to update/upload a photo should look a bit like:

          ```html
          <form action="/:userId/photo" method="POST" enctype="multipart/form-data">
            <input type="file" name="photo">
            <button type="submit">Upload Photo</button>
          </form>
          ```

          NB  input `name` has to be 'photo' `userId` should be the id of the user you want a photo of. 

          Courtesy of ChatGPT. 

          Haven't tested it! - you will need to have a `handleSubmit()` handler to add things such as auth etc.
      tags:
        - User
  '/user/{userId}/password':
    patch:
      summary: Changes a user password
      operationId: patch-user-password
      responses:
        '200':
          description: OK
      description: |-
        Changes the password of a user. 

        Changes the password of a user found in the auth header, once it passes auth - user is allowed to change their own password

        Admins have access to change any password. Otherwise, if it's a user, the id specified in the path param must match the one found in authorization.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  description: String value for the new password
        description: ''
      tags:
        - User
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
        description: Id of the user who's password is to be changed
  /login:
    get:
      summary: Handle Login of user
      tags: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAdmin:
                    type: boolean
                  userId:
                    type: string
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      operationId: get-login
      description: |-
        Basically returns a body whether a user is an admin or not, along with a user Id. If credentials are correct in auth header, then all it does is return body as specified below. 

        However, to handle case where creds are incorrrect (e.g. user has entered wrong username, password combination), then a 401 will be returned. Hence on front end, will need to handle case where request is 401, or if request is 200 - then check if a user is an admin or not.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
  /register:
    post:
      summary: Register a new user
      tags:
        - User
      responses:
        '200':
          description: OK
      operationId: get-user-register
      description: |-
        Registers a new user

        Intended to be used when a user on front end presses 'create account' button
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
        description: 'Just the same as a normal user, without the isAdmin attribute'
    parameters: []
components:
  schemas:
    User:
      title: User
      x-stoplight:
        id: gi5rugf7bx6yb
      type: object
      description: Represents a user within the domain
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        isAdmin:
          type: string
        email:
          type: string
        password:
          type: string
    Event:
      title: Event
      x-stoplight:
        id: jvb3r08ohl0mv
      type: object
      properties:
        eventId:
          type: string
        title:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string
      description: Represents an Event within the domain
  securitySchemes:
    Basic Auth:
      name: Authorization
      type: apiKey
      in: header
      description: Use basic auth
  responses:
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: string
            default: Unauthorized
            example: Unauthorized
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              id:
                type: string
security:
  - Authorization: []